Halve text render time by writing in up-to-4 pixel blocks
Pad font pixel data width to multiple of four pixels to support this

+++ tools/gfxc/font.hpp
@@ -8,7 +8,7 @@
 
 struct font_glyph {
 	uint8_t width;
-	std::vector<uint8_t> data;
+	std::vector<uint16_t> data;
 };
 
 struct font {
